US7752 - Molecules Refactor Removing molecules routing and adjusting layouts for sign up and avatar groups
