+++ src/app/layout/header/header.component.html
@@ -97,8 +97,8 @@
               <div class="mega-menu_list">
                 <p class="font-size-large">Molecules</p>
                 <ul>
-                  <li><a routerLink="/ui/molecules/sign-in" routerLinkActive="active">Sign In</a></li>
-                  <li><a routerLink="/ui/molecules/avatar-groups" routerLinkActive="active">Avatar Groups</a></li>
+                  <li><a routerLink="/ui/sign-in" routerLinkActive="active">Sign In</a></li>
+                  <li><a routerLink="/ui/avatar-groups" routerLinkActive="active">Avatar Groups</a></li>
                 </ul>
               </div>
             </div>
