+++ src/app/ui-components/avatar-groups/avatar-groups.component.spec.ts
