--- src/app/ui-components/molecules/avatar-groups/avatar-groups.component.html
@@ -1,58 +1,0 @@
-<div class="page-header">
-  <h2>Avatar Groups</h2>
-</div>
-
-<p>Add one of the following classes to resize an avatar group.</p>
-
-<ul class="list-inline">
-  <li>
-    <code>.avatar-size-1</code>
-  </li>
-  <li>
-    <code>.avatar-size-2</code>
-  </li>
-  <li>
-    <code>.avatar-size-3</code>
-  </li>
-  <li>
-    <code>.avatar-size-4</code>
-  </li>
-  <li>
-    <code>.avatar-size-5</code>
-  </li>
-  <li>
-    <code>.avatar-size-6</code>
-  </li>
-</ul>
-
-<div class="crds-example">
-  <div class="avatar-group avatar-size-1">
-    <img src="http://placehold.it/100/111/fff" class="img-circle" />
-    <img src="http://placehold.it/100/777/fff" class="img-circle" />
-  </div>
-
-  <div class="avatar-group avatar-size-2">
-    <img src="http://placehold.it/100/111/fff" class="img-circle" />
-    <img src="http://placehold.it/100/777/fff" class="img-circle" />
-  </div>
-
-  <div class="avatar-group avatar-size-3">
-    <img src="http://placehold.it/100/111/fff" class="img-circle" />
-    <img src="http://placehold.it/100/777/fff" class="img-circle" />
-  </div>
-
-  <div class="avatar-group avatar-size-4">
-    <img src="http://placehold.it/100/111/fff" class="img-circle" />
-    <img src="http://placehold.it/100/777/fff" class="img-circle" />
-  </div>
-
-  <div class="avatar-group avatar-size-5">
-    <img src="http://placehold.it/100/111/fff" class="img-circle" />
-    <img src="http://placehold.it/100/777/fff" class="img-circle" />
-  </div>
-
-  <div class="avatar-group avatar-size-6">
-    <img src="http://placehold.it/100/111/fff" class="img-circle" />
-    <img src="http://placehold.it/100/777/fff" class="img-circle" />
-  </div>
-</div>
