--- src/app/ui-components/molecules/jumbotron/jumbotron.component.html
@@ -1,5 +1,0 @@
-<div class="jumbotron">
-  <div class="container">
-    <crds-content-block id="UIMoleculesSectionHeader"></crds-content-block>
-  </div>
-</div>
