--- src/app/ui-components/molecules/jumbotron/jumbotron.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'ddk-molecules-jumbotron',
-  templateUrl: './jumbotron.component.html'
-})
-export class MoleculesJumbotronComponent {}
