--- src/app/ui-components/molecules/molecules.component.html
@@ -1,20 +1,0 @@
-<ddk-molecules-jumbotron></ddk-molecules-jumbotron>
-<div class="container">
-  <div class="component">
-    <aside>
-      <nav>
-        <ul class="list-group">
-          <li routerLinkActive="active" class="list-group-item">
-            <a routerLink="/ui/molecules/sign-in">Sign In</a>
-          </li>
-          <li routerLinkActive="active" class="list-group-item">
-            <a routerLink="/ui/molecules/avatar-groups">Avatar Groups</a>
-          </li>
-        </ul>
-      </nav>
-    </aside>
-    <div>
-      <router-outlet></router-outlet>
-    </div>
-  </div>
-</div>
