--- src/app/ui-components/molecules/molecules.component.spec.ts
@@ -1,10 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-import { ElementRef, Renderer } from '@angular/core';
-import { MoleculesComponent } from './molecules.component';
-
-describe('Component: Molecules', () => {
-  it('should create an instance', () => {
-    let component = new MoleculesComponent();
-    expect(component).toBeTruthy();
-  });
-});
