--- src/app/ui-components/molecules/molecules.component.ts
@@ -1,6 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  templateUrl: './molecules.component.html'
-})
-export class MoleculesComponent {}
