+++ src/app/ui-components/sign-in/sign-in.component.html
@@ -1,0 +1,48 @@
+<div class="jumbotron">
+  <div class="container">
+    <crds-content-block id="UISignInHeader"></crds-content-block>
+  </div>
+</div>
+
+<div class="container">
+
+  <div class="col-md-offset-3">
+  
+    <p>Below you can see the layout for the standard sign-in form on crossroads.net and it's related properties. Implementation details for the form (submission and validation) are not included here; any related business logic should be handled by the application in which it is implemented.</p>
+    <p>For a working example in NG2, look at <a href="https://github.com/crdschurch/crds-embed/blob/development/src/app/authentication/authentication.component.html#L14-L55">crds-embed's</a> authentication component. For an example in NG1.5 see <a href="https://github.com/crdschurch/crds-angular/blob/development/crossroads.net/core/login/login_form.html#L7-L38">crds-angular</a> login form template and controller.</p>
+    <div class="crds-example">
+      <header class="page-header">
+        <h2>Sign In</h2>
+      </header>
+      <form novalidate >
+        <div class="form-group">
+          <label class="sr-only" for="email">Email</label>
+          <input type="email" class="form-control" placeholder="Email" name="email">
+        </div>
+
+        <div class="form-group">
+          <label class="sr-only" for="password">Password</label>
+          <input type="password" class="form-control" placeholder="Password" name="password">
+        </div>
+
+        <div class="row">
+          <div class="col-sm-12">
+            <p class="pull-right text-left text-block">
+              <a href="#" target="_blank">Forgot password?</a><br>
+              <a href="#" target="_blank">Need help?</a>
+            </p>
+          </div>
+        </div>
+
+        <hr>
+
+        <p class="text-block">Or <a href="#" >register for an account</a></p>
+        <footer class="clearfix">
+          <input type="submit" class="btn btn-primary pull-right" value="Sign In">
+        </footer>
+      </form>
+    </div>
+
+
+  </div>
+</div>
